# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Teste Técnico - Desenvolvedor Pleno Golang Onfly - Documentação",
        "title": "Travel API Documentation",
        "contact": {
            "name": "My Github",
            "url": "https://github.com/rodrigoPQF"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1/travel",
    "paths": {
        "/request": {
            "post": {
                "description": "Creates a new travel request with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Travel Request"
                ],
                "summary": "Creates a new travel request",
                "parameters": [
                    {
                        "description": "Travel Request creation payload",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.CreateTravelRequestInputDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    }
                }
            }
        },
        "/request/all": {
            "get": {
                "description": "Fetches all travel requests based on query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Travel Request"
                ],
                "summary": "Retrieves a list of all travel requests",
                "parameters": [
                    {
                        "enum": [
                            "APPROVED",
                            "CANCELED"
                        ],
                        "type": "string",
                        "description": "Status filter",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseResponseHandler"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.TravelRequestDto"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    }
                }
            }
        },
        "/request/{id}": {
            "get": {
                "description": "Fetches details of a travel request by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Travel Request"
                ],
                "summary": "Retrieves a travel request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Travel Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BaseResponseHandler"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/dtos.TravelRequestDto"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates details of a travel request by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Travel Request"
                ],
                "summary": "Updates a travel request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Travel Request ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request body for updating a travel request",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.UpdateTravelRequestBodyInputDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.BaseResponseHandler"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.CreateTravelRequestInputDto": {
            "type": "object",
            "required": [
                "applicantName",
                "departureDate",
                "destination",
                "id",
                "returnDate",
                "status"
            ],
            "properties": {
                "applicantName": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Raul"
                },
                "departureDate": {
                    "type": "string",
                    "example": "2001-01-30"
                },
                "destination": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3,
                    "example": "Rio de Janeiro"
                },
                "id": {
                    "type": "string",
                    "example": "f57e8039-c69d-4de6-b483-eab41c804d16"
                },
                "returnDate": {
                    "type": "string",
                    "example": "2001-06-30"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "REQUESTED",
                        "APPROVED",
                        "CANCELED"
                    ]
                }
            }
        },
        "dtos.TravelRequestDto": {
            "type": "object",
            "properties": {
                "applicantName": {
                    "type": "string",
                    "example": "Raul"
                },
                "departureDate": {
                    "type": "string",
                    "example": "2001-01-30"
                },
                "destination": {
                    "type": "string",
                    "example": "Rio de Janeiro"
                },
                "id": {
                    "type": "string",
                    "example": "f57e8039-c69d-4de6-b483-eab41c804d16"
                },
                "returnDate": {
                    "type": "string",
                    "example": "2002-06-30"
                },
                "status": {
                    "type": "string",
                    "example": "REQUESTED"
                }
            }
        },
        "dtos.UpdateTravelRequestBodyInputDto": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "APPROVED",
                        "CANCELED"
                    ]
                }
            }
        },
        "handlers.BaseResponseHandler": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {},
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}